Collapse Qt in-process channel classes together
