+++ IPython/kernel/inprocess/channels.py
@@ -43,6 +43,9 @@ class InProcessChannel(object):
         """
         raise NotImplementedError('call_handlers must be defined in a subclass.')
 
+    def flush(self, timeout=1.0):
+        pass
+
     #--------------------------------------------------------------------------
     # InProcessChannel interface
     #--------------------------------------------------------------------------
@@ -70,102 +73,14 @@ class InProcessShellChannel(InProcessChannel):
 
     # flag for whether execute requests should be allowed to call raw_input
     allow_stdin = True
-    proxy_methods = [
-        'execute',
-        'complete',
-        'inspect',
-        'history',
-        'shutdown',
-        'kernel_info',
-    ]
-
-    #--------------------------------------------------------------------------
-    # ShellChannel interface
-    #--------------------------------------------------------------------------
-
-    def execute(self, code, silent=False, store_history=True,
-                user_expressions={}, allow_stdin=None):
-        if allow_stdin is None:
-            allow_stdin = self.allow_stdin
-        content = dict(code=code, silent=silent, store_history=store_history,
-                       user_expressions=user_expressions,
-                       allow_stdin=allow_stdin)
-        msg = self.client.session.msg('execute_request', content)
-        self._dispatch_to_kernel(msg)
-        return msg['header']['msg_id']
-
-    def complete(self, code, cursor_pos=None):
-        if cursor_pos is None:
-            cursor_pos = len(code)
-        content = dict(code=code, cursor_pos=cursor_pos)
-        msg = self.client.session.msg('complete_request', content)
-        self._dispatch_to_kernel(msg)
-        return msg['header']['msg_id']
-
-    def inspect(self, code, cursor_pos=None, detail_level=0):
-        if cursor_pos is None:
-            cursor_pos = len(code)
-        content = dict(code=code, cursor_pos=cursor_pos,
-            detail_level=detail_level,
-        )
-        msg = self.client.session.msg('inspect_request', content)
-        self._dispatch_to_kernel(msg)
-        return msg['header']['msg_id']
-
-    def history(self, raw=True, output=False, hist_access_type='range', **kwds):
-        content = dict(raw=raw, output=output,
-                       hist_access_type=hist_access_type, **kwds)
-        msg = self.client.session.msg('history_request', content)
-        self._dispatch_to_kernel(msg)
-        return msg['header']['msg_id']
-
-    def shutdown(self, restart=False):
-        # FIXME: What to do here?
-        raise NotImplementedError('Cannot shutdown in-process kernel')
-
-    def kernel_info(self):
-        """Request kernel info."""
-        msg = self.client.session.msg('kernel_info_request')
-        self._dispatch_to_kernel(msg)
-        return msg['header']['msg_id']
-
-    #--------------------------------------------------------------------------
-    # Protected interface
-    #--------------------------------------------------------------------------
-
-    def _dispatch_to_kernel(self, msg):
-        """ Send a message to the kernel and handle a reply.
-        """
-        kernel = self.client.kernel
-        if kernel is None:
-            raise RuntimeError('Cannot send request. No kernel exists.')
-
-        stream = DummySocket()
-        self.client.session.send(stream, msg)
-        msg_parts = stream.recv_multipart()
-        kernel.dispatch_shell(stream, msg_parts)
-
-        idents, reply_msg = self.client.session.recv(stream, copy=False)
-        self.call_handlers_later(reply_msg)
-
 
 class InProcessIOPubChannel(InProcessChannel):
     """See `IPython.kernel.channels.IOPubChannel` for docstrings."""
-
-    def flush(self, timeout=1.0):
-        pass
 
 
 class InProcessStdInChannel(InProcessChannel):
     """See `IPython.kernel.channels.StdInChannel` for docstrings."""
 
-    proxy_methods = ['input']
-
-    def input(self, string):
-        kernel = self.client.kernel
-        if kernel is None:
-            raise RuntimeError('Cannot send input reply. No kernel exists.')
-        kernel.raw_input_str = string
 
 
 class InProcessHBChannel(InProcessChannel):
